Introducing the repeat handler
The repeat handler aims to allow the designer draw a template
inline, just inside the element which is going to be the
template rendering target. It uses its own template engine, but
doens't change the default one, relying on the KnockoutJS
capability of specifying the engine as an option to the template
handler.