Dependent bindings (actually); SELECT options/value fixes
Changes in bindingAttributeSyntax make it so that a binding that accesses another binding will also be dependent on that binding's dependentObservable. This is used to make the value and options bindings work together properly.  Since changes to options may affect the value, the value binding is now dependent on the options binding.  Code in the options binding that used to be necessary to initilize the SELECT properly if the value binding was specified before options is no longer required since the value binding handler will automatically call the options binding handler first if hasn't been done so already.  There are also changes in selectExtensions.js to enable de-selecting or resetting the selection.