Access to observables in init function shouldn't cause subscription; update build
If you accessed an observable within an init function, and that init function was called within a dependentObservable (as it is if it's in a template), the outer dependentObservable would incorrectly get a subscription to the inner observable. I fixed the immediate problem in the foreach and template bindings which unecessarily unbound an observable. I also wrapped the whole init code in a dependencyDetection block that just ignores dependencies.