Use a JavaScript Object to detect for existing dependencies.
Previously, Knockout was using a Javascript Array. This means that checking
to see if a dependency already existed for a particular subscribable was
an O(n) operation on the lenght of the array. While Chrome and others could
optimize this in their native Array.indexOf calls, browsers that don't support
a native version of that call perform horribly while executing KO's version
on a large dependency set (i.e. IE8).