#460 - Always use value attribute, if specified, even if blank.
Add spec to show that a blank value will be used instead of the option text. Modern browsers automatically set element.value to the text if the value attribute is missing. IE<=7 doesn't and also makes it kind of hard to tell of the value attribute is set or not.