#476 - make 'with' binding more lightweight (also 'if' and 'ifnot')
Key features:
If the value is initially true, will save copy of nodes and bind to the originals;
if the value is initially false, will save the originals.
Will only update UI if the value has changed (for if/ifnot, if the value's truthiness has changed).
Uses cloneNode to copy nodes.