Add a global before and after callback for foreach/template bindings
These global callbacks give you a single opportunity to execute code before
and after the DOM is modified during a mapping of an array to DOM node
children. This allows you to do things like animate the transition between old
and new states in a way that requires knowing the old and new location of
child elements. With the current callbacks, the only callback that fires before
the DOM is updated is beforeMove, so if there aren't any moves there's no way
to execute code to check on old positions before the DOM is changed.