Change makeArray to just wrap arrayPushAll
The code in makeArray and arrayPushAll does almost exactly the same thing, with the caveat that arrayPushAll has an optimization if the object being passed in is already an array.  This just puts everything through the same code path.

Originally part of #1199, separated out since it's just simple refactoring and not really related to the other changes there.