Fix for IE7 spec failure when run against knockout-raw.js
For some reason, on IE7 json2.js is treated as not yet loaded when
knockout-raw.js first starts running, even though its script tag is
higher up in the page. So it doesn’t work to capture a reference to it
in the test. However this doesn’t affect production builds.