Capture jQuery as jQueryInstance to avoid IE < 9 issue
On IE < 9, the line “var jQuery = window.jQuery” overwrites
window.jQuery with undefined even if it was defined before, if executed
in global scope. This didn’t affect the compiled version of KO, but it
meant that when running specs in IE < 9 against knockout-raw.js, jQuery
was always treated as not referenced even if it was referenced.