Move 'pure computed' intro down a bit, and emphasise the use case
Reason: For people totally new to KO, it may seem strange and
off-putting to get the discussion about regular-vs-pure computeds right
at the top of the first doc page about computeds. I think it fits
better a bit further down, when a new reader has a better sense of what
a computed is for.