Enhance template binding to accept a "nodes" array
When injecting supplied nodes into a component output, it’s useful to
be able to pass the node array directly from the component viewmodel
and bind to those nodes in the view. If we didn’t have this feature,
you’d have to dynamically create a named template or something, which
would be awful.