Fix for potential memory leak
We detected that when we remove a component binding from the page (via a knockout if binding) then for some reason our viewModel still lives in the heap. The retainer (in the memory profiler) seems to be the "currentViewModel" variable... I wonder if it is required once the viewModel has been disposed? Because setting it to null like this seems to fix the problem and completely remove the viewModel from the heap. Is there another way this gets cleaned up that we're missing?