Support afterRender for if, ifnot and with bindings
I saw several related issues but realized that it's still not
implemented.

So here is a basic support for the afterRender callback to if, ifnot and
with bindings. One can use it like this

```
with: data, afterRender: callback
if: condition, afterRender: callback
ifnot: condition, afterRender: callback
```

As far as other previously mentioned bindings go (visible, etc.) I
personally decided not to implement it for them because for example
visible is semantically different, there is nothing like "rendering" in
this sense.

Semantics are same as with foreach and template bindings.

Added the some specs too, based on the specs found for foreach binding.