Update isComputed evaluation
When knockout.mapping is used, it replaces ko.computed with its own definition of ko.computed, and ends up causing isComputed to return incorrect results (ie true for computed observables). This patch introduces a new ko.originalComputed variable that should not be overwritten/reassigned, and is used in the evaluation of isComputed.