Fixed leaking when tracking array changes
The observableArray did leak the last value held by the observable when
subscribing for array change events, and then disposing the subscription.

The vlaue was being held by the closure held by the
arrayChangeSubscription. Setting it to null allows the GC to collect the
closure and the previousContents variable.

This commit fixes #2216