Fix side effects on awake callback
Previously if a pure computed in a chain of pure computeds has a subscription
to its awake event which causes side effects that update its value, it was
possible for listeners to get into a bad state. This only happened if the pure
computed in question was invoked before subscribing.